requirements.txt
python-dotenv==1.0.1
langchain==0.1.5
langchain-openai==0.0.5